Cleaned up bell logic, added dynamic joint bell
Bell logic had a few bugs due to type checking with IsCharacter not being on every type of collider   that the bell could interact with. Fixed that, Made the trigger hurtbox added on load so that the layer could be correct, filled out a bug report for apg and ahss since they are not triggering OnHit events. I also made sure that the network event is only sent if the object sending is owned by the current user to avoid double sending hit events.

Added a dynamic version of the bell that has physics using a rigidbody and joint. The moving bell looks and works pretty good in practice but gets a bit buggy when interacted with by titans so I removed the trigger for non-character types.