Add HoverTooltip
Generalize interface of TooltipPopup to allow for callers other than Buttons.
HoverTooltip shows and hides the popup when the mouse enters and exits a graphic.
Can be added seamlessly through code, or in the editor.
HoverTooltips added through code can be customized by adding in the editor instead and modifying the default values of the properties, assuming the code doesn't set them.
This is accomplished through GetOrAddComponent. Make sure to use it.

Refs: https://discord.com/channels/681641241125060652/1284582772769292298