Resetting Escape Ability
Added ReduceCooldownAmount to BaseUsable's constructor default value is 0f. It can be used if we want to reduce cooldown based on any condition. It's only used in EscapeSpecial for now.  Also added HasDurability to BaseUsable and than override this in both AmmoWeapon and BladeWeapon. Main reason i did this is BladeWeapon.CanUse() is also looking for HumanState.Idle and this disables blade weapon to use escape special. Escape Special has 300s cooldown, 50f ReduceCooldownAmount and its unable to use at start.