Top Left Hud finished, other optimizations
Top left hud now split into two nested sub-canvases which fixes telemetry forcing redraws on the playerlist. Botton hud, chat, and everything else still forces redraws of the nested canvas's but its whatever at this point.

Fixed a gc alloc issue in weapon trail as it was creating a new list every frame for all points in the trail. (also earlier made it so the trail can be disabled, set to only the main players', or all players)

Added a dark transparent background behind the players name if highly visible names are enabled.

Fixed a bug with the ping property where an exception would be thrown if the property was not set yet.

Removed RPC time sync via photon timestamps as it gave an inaccurate timestamp once in a server I was in and I can't figure out what caused the error case.