Added back in arraypool
Might be worthwhile to write a custom arraypool since unity's doesn't seem to consistently avoid allocs.
What I am looking for is reserving a somewhat fixed amount of memory for cl and calling it a day since most performant-required gamemodes shouldn't exceed 1mb of memory, we could allocate towards that amount with our pool and likely never have to allocate temp objects afterwards. - same for commonly used objects (Vector3, Quaternion, List, Dict) -> we can probably get away with caching a fixed count of these by counting the amount alloc'd on frame or on tick. This would be simpler than adding in generational gc/pool but would get most of the benefits out of the way.