:/ ig two calls is fine.
Networked objects work for now but its not optimal to send the entire configuration over every time.

Need to enforce a way for host to define prefabs and then to reuse them, otherwise the current custom sync methodology works better.

Objects appear to be cleaned up properly now when destroyed.

Networked objects are immediately created and returned for the creator.

No methodology for finding networked objects on the client side exists besides pre-loading a component that signals.