Added support for unary operators
added unary +/- and binop %
Digits are no longer preparsed as negative and now use Unary(-, expr) as a standard.

This should solve the bug we were seeing with
```
a = 10;
b = -a;
```
throwing an error.
This should also fix the issue we were seeing with
```
f =  5 -2;
```
returning -2 instead of actually evaluating the expression.